Create bounceless_ngon_puzzle.py
Finds the solution to another weird puzzle I came up with;

A particle lies at a vertex on the interior of a regular n-gon with 1m long sides, and is launched perfectly parallel to one of the walls directly adjacent to it and moves along it at 1ms^-1. The walls are perfectly inelastic / have coefficient of restitution 0. For which n does the particle complete one full rotation of the shape the fastest?

Quite interesting! It's obvious that for very small n values, you take ages because the interior angles are so sharp that you lose much of your velocity when you hit it - then as you increase n, the time gets faster and faster since the angles are less sharp, but then it begins to increase again, eventually asymptotically going towards the line y=x, since for incredibly shallow angles, you lose almost no speed, so it would just be the perimiter / speed.